# docs/swagger.yaml
components:
  schemas:
...
    BaseResponse:
      properties:
        msg:
          default: ''
          title: Msg
          type: string
        status:
          default: success
          title: Status
          type: string
      title: BaseResponse
      type: object
...
    CreateDocument:
      properties:
        parent_document_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Parent Document Id
        title:
          title: Title
          type: string
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - title
      - workspace_id
      title: CreateDocument
      type: object
...
    CreateWorkspace:
      properties:
        title:
          title: Title
          type: string
      required:
      - title
      title: CreateWorkspace
      type: object
...
    DocumentInfoResponse:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        creator_user:
          $ref: '#/components/schemas/UserBaseInfoResponse'
        id:
          format: uuid
          title: Id
          type: string
        last_published_version_at:
          format: date-time
          title: Last Published Version At
          type: string
        parent_document_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Parent Document Id
        permission_mode:
          $ref: '#/components/schemas/ObjectPermissionMode'
        title:
          title: Title
          type: string
        workspace_id:
          format: uuid
          title: Workspace Id
          type: string
      required:
      - permission_mode
      - id
      - title
      - workspace_id
      - creator_user
      - created_at
      - last_published_version_at
      title: DocumentInfoResponse
      type: object
...
    DocumentTemplateInfoResponse:
      properties:
        creator_user_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Creator User Id
        description:
          title: Description
          type: string
        document_id:
          format: uuid
          title: Document Id
          type: string
        document_template_type:
          allOf:
          - $ref: '#/components/schemas/DocumentTemplateType'
          default: START
        id:
          format: uuid
          title: Id
          type: string
        title:
          title: Title
          type: string
      required:
      - id
      - title
      - description
      - document_id
      title: DocumentTemplateInfoResponse
      type: object
    DocumentTemplateType:
      const: START
      title: DocumentTemplateType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ObjectPermissionMode:
      enum:
      - HIDDEN_INACCESSIBLE
      - INACCESSIBLE
      - VIEW_AND_EXPORT
      - ABILITY_USE_TEMPLATE
      - EDITING
      - DELETION
      title: ObjectPermissionMode
      type: string
...
    Page_DocumentTemplateInfoResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DocumentTemplateInfoResponse'
          title: Items
          type: array
        page:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          title: Page
        pages:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Pages
        size:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          title: Size
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - size
      title: Page[DocumentTemplateInfoResponse]
      type: object
...
    ResponsibilityType:
      enum:
      - VIEWER
      - EDITOR
      - ADMIN
      title: ResponsibilityType
      type: string
...
    UserBaseInfoResponse:
      properties:
        email:
          format: email
          title: Email
          type: string
        first_name:
          title: First Name
          type: string
        last_name:
          title: Last Name
          type: string
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        second_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Second Name
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        wiki_api_client:
          anyOf:
          - $ref: '#/components/schemas/WikiApiClientInfoResponse'
          - type: 'null'
      required:
      - email
      - first_name
      - last_name
      title: UserBaseInfoResponse
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    WikiApiClientInfoResponse:
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          format: uuid
          title: Id
          type: string
        is_enabled:
          default: true
          title: Is Enabled
          type: boolean
        responsibility:
          allOf:
          - $ref: '#/components/schemas/ResponsibilityType'
          default: VIEWER
      required:
      - id
      title: WikiApiClientInfoResponse
      type: object
...
    WorkspaceInfoResponse:
      properties:
        id:
          format: uuid
          title: Id
          type: string
        owner_user:
          $ref: '#/components/schemas/UserBaseInfoResponse'
        permission_mode:
          $ref: '#/components/schemas/ObjectPermissionMode'
        title:
          title: Title
          type: string
      required:
      - permission_mode
      - id
      - title
      - owner_user
      title: WorkspaceInfoResponse
      type: object
  securitySchemes:
    WikiAccessToken:
      in: cookie
      name: wiki_access_token
      type: apiKey
    WikiApiKey:
      in: header
      name: X-Wiki-API-Key
      type: apiKey
    WikiBearer:
      scheme: bearer
      type: http
    WikiRootApiKey:
      description: 'Root key, allows you to perform actions as administrator, random
        user data is generated.

        **Limitation:** operations requiring an explicitly registered userare unavailable,
        example.: `Create: POST /api_key/`.'
      in: header
      name: X-Wiki-ROOT-API-Key
      type: apiKey
info:
  description: Welcome to Wiki's API documentation! Here you will able to discover
    all of the ways you can interact with the Wiki API.
  title: Wiki API
  version: 0.0.1
openapi: 3.1.0
paths:
...
  /api/v1/document/:
    post:
      description: '## Create document

        Create a document using a template,

        if template_id=None then will be created empty document,

        if parent_document_id=None then document is in the root.'
      operationId: create_document_api_v1_document__post
      parameters:
      - in: query
        name: template_id
        required: false
        schema:
          format: uuid
          title: Template Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocument'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Create document
      tags:
      - WorkspaceDocument
  /api/v1/document/all:
    get:
      operationId: get_documents_by_workspace_id_api_v1_document_all_get
      parameters:
      - in: query
        name: workspace_id
        required: true
        schema:
          format: uuid
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DocumentInfoResponse'
                title: Response Get Documents By Workspace Id Api V1 Document All
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all document by workspace id
      tags:
      - WorkspaceDocument
...
  /api/v1/document/{document_id}:
    delete:
      description: '## Deleted document by id

        **You have to have a deletion permission or be an administrator.**


        When a document is deleted, the hierarchy changes:

        all child documents become child documents of the parent.'
      operationId: delete_document_api_v1_document__document_id__delete
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Delete document
      tags:
      - WorkspaceDocument
...
  /api/v1/document_template/:
    get:
      operationId: get_document_template_by_id_api_v1_document_template__get
      parameters:
      - in: query
        name: document_template_id
        required: true
        schema:
          format: uuid
          title: Document Template Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTemplateInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add document template
      tags:
      - DocumentTemplate
    post:
      operationId: create_document_template_api_v1_document_template__post
      parameters:
      - in: query
        name: title
        required: true
        schema:
          title: Title
          type: string
      - in: query
        name: description
        required: true
        schema:
          title: Description
          type: string
      - in: query
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: query
        name: orig_commit_id
        required: true
        schema:
          title: Orig Commit Id
          type: string
      - in: query
        name: document_template_type
        required: false
        schema:
          allOf:
          - $ref: '#/components/schemas/DocumentTemplateType'
          default: START
          title: Document Template Type
      - in: query
        name: creator_user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Creator User Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTemplateInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Add template
      tags:
      - DocumentTemplate
  /api/v1/document_template/all:
    get:
      operationId: get_all_document_templates_api_v1_document_template_all_get
      parameters:
      - in: query
        name: title
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      - in: query
        name: description
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      - in: query
        name: document_template_type
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DocumentTemplateType'
          - type: 'null'
          title: Document Template Type
      - in: query
        name: creator_user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Creator User Id
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_DocumentTemplateInfoResponse_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all document templates
      tags:
      - DocumentTemplate
...
  /api/v1/workspace/:
    post:
      operationId: create_workspace_api_v1_workspace__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspace'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Create workspace
      tags:
      - Workspace
  /api/v1/workspace/all:
    get:
      operationId: get_workspaces_api_v1_workspace_all_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkspaceInfoResponse'
                title: Response Get Workspaces Api V1 Workspace All Get
                type: array
          description: Successful Response
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all workspace
      tags:
      - Workspace
  /api/v1/workspace/info:
    get:
      operationId: get_workspace_info_api_v1_workspace_info_get
      parameters:
      - in: query
        name: workspace_id
        required: true
        schema:
          format: uuid
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get workspace info
      tags:
      - Workspace
...
servers:
- description: Local server
  url: http://localhost:8000/
- description: Demo server
  url: https://api.wiki.example.com/
- description: Ngrok free server
  url: https://poorly-ideal-cobra.ngrok-free.app/